+++ pkg/oas/oas_parameters_gen.go
@@ -4294,6 +4294,72 @@ func decodeGetBlockchainBlockTransactionsParams(args [1]string, argsEscaped bool
 	return params, nil
 }
 
+// GetBlockchainConfigFromBlockParams is parameters of getBlockchainConfigFromBlock operation.
+type GetBlockchainConfigFromBlockParams struct {
+	// Masterchain block seqno.
+	MasterchainSeqno int32
+}
+
+func unpackGetBlockchainConfigFromBlockParams(packed middleware.Parameters) (params GetBlockchainConfigFromBlockParams) {
+	{
+		key := middleware.ParameterKey{
+			Name: "masterchain_seqno",
+			In:   "path",
+		}
+		params.MasterchainSeqno = packed[key].(int32)
+	}
+	return params
+}
+
+func decodeGetBlockchainConfigFromBlockParams(args [1]string, argsEscaped bool, r *http.Request) (params GetBlockchainConfigFromBlockParams, _ error) {
+	// Decode path: masterchain_seqno.
+	if err := func() error {
+		param := args[0]
+		if argsEscaped {
+			unescaped, err := url.PathUnescape(args[0])
+			if err != nil {
+				return errors.Wrap(err, "unescape path")
+			}
+			param = unescaped
+		}
+		if len(param) > 0 {
+			d := uri.NewPathDecoder(uri.PathDecoderConfig{
+				Param:   "masterchain_seqno",
+				Value:   param,
+				Style:   uri.PathStyleSimple,
+				Explode: false,
+			})
+
+			if err := func() error {
+				val, err := d.DecodeValue()
+				if err != nil {
+					return err
+				}
+
+				c, err := conv.ToInt32(val)
+				if err != nil {
+					return err
+				}
+
+				params.MasterchainSeqno = c
+				return nil
+			}(); err != nil {
+				return err
+			}
+		} else {
+			return validate.ErrFieldRequired
+		}
+		return nil
+	}(); err != nil {
+		return params, &ogenerrors.DecodeParamError{
+			Name: "masterchain_seqno",
+			In:   "path",
+			Err:  err,
+		}
+	}
+	return params, nil
+}
+
 // GetBlockchainMasterchainShardsParams is parameters of getBlockchainMasterchainShards operation.
 type GetBlockchainMasterchainShardsParams struct {
 	// Masterchain block seqno.
@@ -7005,23 +7071,23 @@ func decodeGetRawBlockchainBlockStateParams(args [1]string, argsEscaped bool, r 
 
 // GetRawBlockchainConfigFromBlockParams is parameters of getRawBlockchainConfigFromBlock operation.
 type GetRawBlockchainConfigFromBlockParams struct {
-	// Block ID.
-	BlockID string
+	// Masterchain block seqno.
+	MasterchainSeqno int32
 }
 
 func unpackGetRawBlockchainConfigFromBlockParams(packed middleware.Parameters) (params GetRawBlockchainConfigFromBlockParams) {
 	{
 		key := middleware.ParameterKey{
-			Name: "block_id",
+			Name: "masterchain_seqno",
 			In:   "path",
 		}
-		params.BlockID = packed[key].(string)
+		params.MasterchainSeqno = packed[key].(int32)
 	}
 	return params
 }
 
 func decodeGetRawBlockchainConfigFromBlockParams(args [1]string, argsEscaped bool, r *http.Request) (params GetRawBlockchainConfigFromBlockParams, _ error) {
-	// Decode path: block_id.
+	// Decode path: masterchain_seqno.
 	if err := func() error {
 		param := args[0]
 		if argsEscaped {
@@ -7033,7 +7099,7 @@ func decodeGetRawBlockchainConfigFromBlockParams(args [1]string, argsEscaped boo
 		}
 		if len(param) > 0 {
 			d := uri.NewPathDecoder(uri.PathDecoderConfig{
-				Param:   "block_id",
+				Param:   "masterchain_seqno",
 				Value:   param,
 				Style:   uri.PathStyleSimple,
 				Explode: false,
@@ -7045,12 +7111,12 @@ func decodeGetRawBlockchainConfigFromBlockParams(args [1]string, argsEscaped boo
 					return err
 				}
 
-				c, err := conv.ToString(val)
+				c, err := conv.ToInt32(val)
 				if err != nil {
 					return err
 				}
 
-				params.BlockID = c
+				params.MasterchainSeqno = c
 				return nil
 			}(); err != nil {
 				return err
@@ -7061,7 +7127,7 @@ func decodeGetRawBlockchainConfigFromBlockParams(args [1]string, argsEscaped boo
 		return nil
 	}(); err != nil {
 		return params, &ogenerrors.DecodeParamError{
-			Name: "block_id",
+			Name: "masterchain_seqno",
 			In:   "path",
 			Err:  err,
 		}