+++ tonapi/oas_client_gen.go
@@ -219,6 +219,12 @@ type Invoker interface {
 	//
 	// GET /v2/blockchain/config
 	GetBlockchainConfig(ctx context.Context) (*BlockchainConfig, error)
+	// GetBlockchainConfigFromBlock invokes getBlockchainConfigFromBlock operation.
+	//
+	// Get blockchain config from a specific block, if present.
+	//
+	// GET /v2/blockchain/masterchain/{masterchain_seqno}/config
+	GetBlockchainConfigFromBlock(ctx context.Context, params GetBlockchainConfigFromBlockParams) (*BlockchainConfig, error)
 	// GetBlockchainMasterchainHead invokes getBlockchainMasterchainHead operation.
 	//
 	// Get last known masterchain block.
@@ -390,7 +396,7 @@ type Invoker interface {
 	//
 	// Get raw blockchain config from a specific block, if present.
 	//
-	// GET /v2/blockchain/blocks/{block_id}/config/raw
+	// GET /v2/blockchain/masterchain/{masterchain_seqno}/config/raw
 	GetRawBlockchainConfigFromBlock(ctx context.Context, params GetRawBlockchainConfigFromBlockParams) (*RawBlockchainConfig, error)
 	// GetRawConfig invokes getRawConfig operation.
 	//
@@ -4218,6 +4224,97 @@ func (c *Client) sendGetBlockchainConfig(ctx context.Context) (res *BlockchainCo
 	return result, nil
 }
 
+// GetBlockchainConfigFromBlock invokes getBlockchainConfigFromBlock operation.
+//
+// Get blockchain config from a specific block, if present.
+//
+// GET /v2/blockchain/masterchain/{masterchain_seqno}/config
+func (c *Client) GetBlockchainConfigFromBlock(ctx context.Context, params GetBlockchainConfigFromBlockParams) (*BlockchainConfig, error) {
+	res, err := c.sendGetBlockchainConfigFromBlock(ctx, params)
+	return res, err
+}
+
+func (c *Client) sendGetBlockchainConfigFromBlock(ctx context.Context, params GetBlockchainConfigFromBlockParams) (res *BlockchainConfig, err error) {
+	otelAttrs := []attribute.KeyValue{
+		otelogen.OperationID("getBlockchainConfigFromBlock"),
+		semconv.HTTPMethodKey.String("GET"),
+		semconv.HTTPRouteKey.String("/v2/blockchain/masterchain/{masterchain_seqno}/config"),
+	}
+
+	// Run stopwatch.
+	startTime := time.Now()
+	defer func() {
+		// Use floating point division here for higher precision (instead of Millisecond method).
+		elapsedDuration := time.Since(startTime)
+		c.duration.Record(ctx, float64(float64(elapsedDuration)/float64(time.Millisecond)), metric.WithAttributes(otelAttrs...))
+	}()
+
+	// Increment request counter.
+	c.requests.Add(ctx, 1, metric.WithAttributes(otelAttrs...))
+
+	// Start a span for this request.
+	ctx, span := c.cfg.Tracer.Start(ctx, "GetBlockchainConfigFromBlock",
+		trace.WithAttributes(otelAttrs...),
+		clientSpanKind,
+	)
+	// Track stage for error reporting.
+	var stage string
+	defer func() {
+		if err != nil {
+			span.RecordError(err)
+			span.SetStatus(codes.Error, stage)
+			c.errors.Add(ctx, 1, metric.WithAttributes(otelAttrs...))
+		}
+		span.End()
+	}()
+
+	stage = "BuildURL"
+	u := uri.Clone(c.requestURL(ctx))
+	var pathParts [3]string
+	pathParts[0] = "/v2/blockchain/masterchain/"
+	{
+		// Encode "masterchain_seqno" parameter.
+		e := uri.NewPathEncoder(uri.PathEncoderConfig{
+			Param:   "masterchain_seqno",
+			Style:   uri.PathStyleSimple,
+			Explode: false,
+		})
+		if err := func() error {
+			return e.EncodeValue(conv.Int32ToString(params.MasterchainSeqno))
+		}(); err != nil {
+			return res, errors.Wrap(err, "encode path")
+		}
+		encoded, err := e.Result()
+		if err != nil {
+			return res, errors.Wrap(err, "encode path")
+		}
+		pathParts[1] = encoded
+	}
+	pathParts[2] = "/config"
+	uri.AddPathParts(u, pathParts[:]...)
+
+	stage = "EncodeRequest"
+	r, err := ht.NewRequest(ctx, "GET", u)
+	if err != nil {
+		return res, errors.Wrap(err, "create request")
+	}
+
+	stage = "SendRequest"
+	resp, err := c.cfg.Client.Do(r)
+	if err != nil {
+		return res, errors.Wrap(err, "do request")
+	}
+	defer resp.Body.Close()
+
+	stage = "DecodeResponse"
+	result, err := decodeGetBlockchainConfigFromBlockResponse(resp)
+	if err != nil {
+		return res, errors.Wrap(err, "decode response")
+	}
+
+	return result, nil
+}
+
 // GetBlockchainMasterchainHead invokes getBlockchainMasterchainHead operation.
 //
 // Get last known masterchain block.
@@ -6982,7 +7079,7 @@ func (c *Client) sendGetRawBlockchainConfig(ctx context.Context) (res *RawBlockc
 //
 // Get raw blockchain config from a specific block, if present.
 //
-// GET /v2/blockchain/blocks/{block_id}/config/raw
+// GET /v2/blockchain/masterchain/{masterchain_seqno}/config/raw
 func (c *Client) GetRawBlockchainConfigFromBlock(ctx context.Context, params GetRawBlockchainConfigFromBlockParams) (*RawBlockchainConfig, error) {
 	res, err := c.sendGetRawBlockchainConfigFromBlock(ctx, params)
 	return res, err
@@ -6992,7 +7089,7 @@ func (c *Client) sendGetRawBlockchainConfigFromBlock(ctx context.Context, params
 	otelAttrs := []attribute.KeyValue{
 		otelogen.OperationID("getRawBlockchainConfigFromBlock"),
 		semconv.HTTPMethodKey.String("GET"),
-		semconv.HTTPRouteKey.String("/v2/blockchain/blocks/{block_id}/config/raw"),
+		semconv.HTTPRouteKey.String("/v2/blockchain/masterchain/{masterchain_seqno}/config/raw"),
 	}
 
 	// Run stopwatch.
@@ -7025,16 +7122,16 @@ func (c *Client) sendGetRawBlockchainConfigFromBlock(ctx context.Context, params
 	stage = "BuildURL"
 	u := uri.Clone(c.requestURL(ctx))
 	var pathParts [3]string
-	pathParts[0] = "/v2/blockchain/blocks/"
+	pathParts[0] = "/v2/blockchain/masterchain/"
 	{
-		// Encode "block_id" parameter.
+		// Encode "masterchain_seqno" parameter.
 		e := uri.NewPathEncoder(uri.PathEncoderConfig{
-			Param:   "block_id",
+			Param:   "masterchain_seqno",
 			Style:   uri.PathStyleSimple,
 			Explode: false,
 		})
 		if err := func() error {
-			return e.EncodeValue(conv.StringToString(params.BlockID))
+			return e.EncodeValue(conv.Int32ToString(params.MasterchainSeqno))
 		}(); err != nil {
 			return res, errors.Wrap(err, "encode path")
 		}