+++ tonapi/oas_parameters_gen.go
@@ -232,6 +232,12 @@ type GetBlockchainBlockTransactionsParams struct {
 	BlockID string
 }
 
+// GetBlockchainConfigFromBlockParams is parameters of getBlockchainConfigFromBlock operation.
+type GetBlockchainConfigFromBlockParams struct {
+	// Masterchain block seqno.
+	MasterchainSeqno int32
+}
+
 // GetBlockchainMasterchainShardsParams is parameters of getBlockchainMasterchainShards operation.
 type GetBlockchainMasterchainShardsParams struct {
 	// Masterchain block seqno.
@@ -397,8 +403,8 @@ type GetRawBlockchainBlockStateParams struct {
 
 // GetRawBlockchainConfigFromBlockParams is parameters of getRawBlockchainConfigFromBlock operation.
 type GetRawBlockchainConfigFromBlockParams struct {
-	// Block ID.
-	BlockID string
+	// Masterchain block seqno.
+	MasterchainSeqno int32
 }
 
 // GetRawConfigParams is parameters of getRawConfig operation.