+++ CHANGELOG.md
@@ -24,3 +24,7 @@
 
 ### 2023-10-01
 - Tonapi sdk
+
+### 2023-10-21
+- `/v2/blockchain/masterchain/{masterchain_seqno}/config` method which returns a blockchain config for a specific block.
+- `/v2/blockchain/masterchain/{masterchain_seqno}/config/raw` method which returns a raw blockchain config for a specific block.